various clean up
