+++ airbyte-integrations/base-java/src/main/java/io/airbyte/integrations/base/IntegrationConfig.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package io.airbyte.integrations.javabase;
+package io.airbyte.integrations.base;
 
 import com.google.common.base.Preconditions;
