+++ airbyte-integrations/base/base.sh
@@ -14,7 +14,7 @@ function error() {
 # todo: make it easy to select source or destination and validate based on selection by adding an integration type env variable.
 function main() {
   CMD="$1"
-  shift 1 || error "command not found."
+  shift 1 || error "command not specified."
 
   ARGS=
   while [ $# -ne 0 ]; do
