+++ airbyte-integrations/csv/Dockerfile
@@ -1,5 +1,4 @@
 FROM airbyte/base-java:dev
-COPY --from=openjdk:14.0.2-slim / /
 
 WORKDIR /airbyte
 # todo (cgardens) - naming.
@@ -7,5 +6,5 @@ ENV INTEGRATION_NAME csv
 
 COPY build/libs/${INTEGRATION_NAME}*.jar lib/${INTEGRATION_NAME}.jar
 
-ENV DESTINATION_CLASS io.airbyte.integrations.csv.CsvDestination
+ENV DESTINATION_CLASS CsvDestination
 ENV DESTINATION_JAR_PATH lib/${INTEGRATION_NAME}.jar
